refactor(features): moved getSubjectDetails to features
